+++ apps/api/src/app/auth/sign-up.service.ts
